massive update for modals and buttons and update on bus assignment table
