+++ components/modal/Assign-Driver/AssignDriverModal.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import Image from 'next/image';
 import Button from "@/components/ui/Button";
 import SearchBar from "@/components/ui/SearchBar";
-import DropdownButton from '../ui/DropdownButton';
+import DropdownButton from '../../ui/DropdownButton';
 import { useEffect} from 'react';
 import { fetchDriversWithToken } from '@/lib/apiCalls/external';
 //import { Driver } from "@/app/interface";
