+++ components/modal/Assign-Route/AssignRouteModal.tsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from 'react';
 import Image from 'next/image';
 import Button from "@/components/ui/Button";
 import SearchBar from "@/components/ui/SearchBar";
-import DropdownButton from '../ui/DropdownButton';
+import DropdownButton from '@/components/ui/DropdownButton';
 import { Route } from '@/app/interface'; // Importing the Route interface
 import { fetchRoutesModalWithToken } from '@/lib/apiCalls/route';
