+++ components/modal/Edit-Route/edit-route.module.css
@@ -1,0 +1,381 @@
+/* Modal Overlay and Base */
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1050;
+  animation: fadeIn 0.3s ease-out;
+}
+
+.modalTitle{
+  display: flex;                  /* Flex container */
+  align-items: center;            /* Center vertically */
+  gap: 0.5rem;                   /* Space between icon/text */
+  font-size: 24px;               /* Large font size */
+  font-weight: 800;              /* Bold */
+  justify-content: center;       /* Center horizontally */
+  color: #ffffff;                /* White text */
+  background-color: #961c1e;     /* Dark red background */
+  border: 4px solid #FFB71F;     /* Yellow border */
+  border-radius: 12px;           /* Rounded corners */
+  padding: 0.75rem 1.5rem;       /* Padding inside */
+  font-family: 'Open Sans', sans-serif; /* Bold font */
+  margin-bottom: 20px;            /* Space below title */
+  box-shadow: 0 4px 6px rgba(0,0,0,0.2); /* Shadow for depth */
+}
+
+.modalDialog {
+  width: 90%;
+  max-width: 800px;
+  max-height: 90vh;
+  overflow-y: auto;
+}
+
+.modalContent {
+  border: none;
+  border-radius: 8px;
+  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.2);
+  background: white;
+  animation: slideUp 0.3s ease-out;
+  display: flex;
+  flex-direction: column;
+}
+
+/* Modal Header */
+.modalHeader {
+  background: #961c1e;
+  color: white;
+  border: none;
+  padding: 20px 25px;
+  border-radius: 8px 8px 0 0;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.modalTitle {
+  font-size: 1.4rem;
+  font-weight: 600;
+  margin: 0;
+}
+
+.closeBtn {
+  background: none;
+  border: none;
+  color: white;
+  font-size: 1.5rem;
+  line-height: 1;
+  opacity: 0.8;
+  transition: opacity 0.2s ease;
+  padding: 0;
+  width: 24px;
+  height: 24px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+}
+
+.closeBtn:hover {
+  opacity: 1;
+  color: white;
+}
+
+/* Modal Body */
+.modalBody {
+  padding: 25px;
+  background: white;
+  flex: 1;
+}
+
+.formSection {
+  background: #f8f9fa;
+  border-radius: 6px;
+  padding: 20px;
+  margin-bottom: 20px;
+  border: 1px solid #e9ecef;
+}
+
+.sectionTitle {
+  font-size: 1rem;
+  font-weight: 600;
+  color: #495057;
+  margin-bottom: 15px;
+  padding-bottom: 8px;
+  border-bottom: 1px solid #dee2e6;
+}
+
+/* Form Elements */
+.row {
+  display: flex;
+  flex-wrap: wrap;
+  margin: -8px;
+}
+
+.col {
+  flex: 1;
+  padding: 8px;
+  min-width: 200px;
+}
+
+.formGroup {
+  margin-bottom: 16px;
+}
+
+.formLabel {
+  font-weight: 500;
+  color: #495057;
+  font-size: 0.9rem;
+  margin-bottom: 6px;
+  display: block;
+}
+
+.formControl {
+  width: 100%;
+  border: 1px solid #ced4da;
+  border-radius: 4px;
+  padding: 10px 12px;
+  font-size: 0.95rem;
+  transition: border-color 0.2s ease;
+  box-sizing: border-box;
+}
+
+.formControl:focus {
+  border-color: #961c1e;
+  box-shadow: 0 0 0 0.2rem rgba(150, 28, 30, 0.25);
+  outline: none;
+}
+
+.selectionInput {
+  cursor: pointer;
+  background: white;
+}
+
+.selectionInput:hover {
+  border-color: #961c1e;
+}
+
+.selectionInput.filled {
+  background: #f8f9fa;
+  border-color: #6c757d;
+}
+
+.regularInput {
+  background: white;
+}
+
+/* Stops Container */
+.stopsScrollContainer {
+  max-height: 300px;
+  overflow-y: auto;
+  padding: 10px;
+  background: #f8f9fa;
+  border-radius: 6px;
+  border: 1px solid #e9ecef;
+}
+
+.stopItem {
+  background: white;
+  border: 1px solid #dee2e6;
+  border-radius: 4px;
+  padding: 10px;
+  margin-bottom: 8px;
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  transition: box-shadow 0.2s ease;
+}
+
+.stopItem:hover {
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+}
+
+.dragHandle {
+  color: #6c757d;
+  cursor: grab;
+  font-size: 1.2rem;
+  padding: 5px;
+  user-select: none;
+}
+
+.dragHandle:active {
+  cursor: grabbing;
+}
+
+.stopInput {
+  flex: 1;
+}
+
+.emptyStopsMessage {
+  color: #6c757d;
+  font-style: italic;
+  text-align: center;
+  padding: 20px;
+  background: white;
+  border-radius: 4px;
+  border: 1px dashed #dee2e6;
+}
+
+/* Buttons */
+.btn {
+  padding: 8px 16px;
+  border-radius: 4px;
+  font-weight: 500;
+  transition: all 0.2s ease;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  cursor: pointer;
+  border: none;
+  text-decoration: none;
+  font-size: 0.9rem;
+}
+
+.btnAddStop {
+  background: #28a745;
+  border: 1px solid #28a745;
+  color: white;
+  margin-top: 12px;
+}
+
+.btnAddStop:hover {
+  background: #218838;
+  border-color: #218838;
+  color: white;
+}
+
+.btnRemoveStop {
+  background: #dc3545;
+  border: 1px solid #dc3545;
+  color: white;
+  padding: 6px 10px;
+  min-width: auto;
+}
+
+.btnRemoveStop:hover {
+  background: #c82333;
+  border-color: #c82333;
+  color: white;
+}
+
+.btnCancel {
+  background: #961C1E;
+  border: 1px solid #FFB71F;
+  padding: 8px 20px;
+  border-radius: 4px;
+  font-weight: 500;
+  color: white;
+  transition: background-color 0.2s ease;
+}
+
+.btnCancel:hover {
+  background: #db1212;
+  border-color: #5a6268;
+  color: white;
+}
+
+.btnSave {
+  background: #961C1E;
+  border: 1px solid #FFB71F;
+  padding: 8px 20px;
+  border-radius: 4px;
+  font-weight: 500;
+  color: white;
+  transition: background-color 0.2s ease;
+}
+
+.btnSave:hover:not(:disabled) {
+  background: #13CE66;
+  border-color: #FFB71F;
+  color: white;
+}
+
+.btnSave:disabled {
+  background: #e9ecef;
+  border-color: #e9ecef;
+  color: #6c757d;
+  cursor: not-allowed;
+}
+
+/* Modal Footer */
+.modalFooter {
+  background: #f8f9fa;
+  border: none;
+  padding: 15px 25px;
+  display: flex;
+  gap: 10px;
+  justify-content: flex-end;
+  border-radius: 0 0 8px 8px;
+}
+
+/* Utilities */
+.iconSmall {
+  width: 16px;
+  height: 16px;
+}
+
+.flexGrow1 {
+  flex: 1;
+}
+
+.mt3 {
+  margin-top: 12px;
+}
+
+.mb3 {
+  margin-bottom: 12px;
+}
+
+/* Animations */
+@keyframes fadeIn {
+  from { opacity: 0; }
+  to { opacity: 1; }
+}
+
+@keyframes slideUp {
+  from { 
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  to { 
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+/* Responsive Design */
+@media (max-width: 768px) {
+  .row {
+    flex-direction: column;
+  }
+  
+  .col {
+    min-width: unset;
+  }
+  
+  .modalDialog {
+    width: 95%;
+    margin: 20px;
+  }
+  
+  .modalHeader,
+  .modalBody,
+  .modalFooter {
+    padding: 15px;
+  }
+  
+  .modalFooter {
+    flex-direction: column;
+  }
+  
+  .modalFooter .btn {
+    width: 100%;
+  }
+}
